Re-enable PRNG seed setting and switch to MT19937
This patch re-enables PRNG seed setting to allow deterministic runs.
Not setting a seed on the command-line will use the current default
behavior of picking a seed from std::random_device. A small refactor
was required to share the PRNG and PRND between the utility and scat3
code so that only one engine needed to be seeded.

This patch switches the PRNG back to Marsenne Twister and utilizes
std::mt19937 from the C++11 STL and removes the vendored mt19937ar code.
std::default_random_engine is implementation-defined and on libstdc++
(used by g++ and icc) it is defined as std::minstd_rand0[^1] which is a
Linear Congruential Generator (LCG) and not suitable for large-scale
MC simulations due to limited (32-bit) state size and modulus[^2][^3].

Future patches should look at other PRNG implementations to provide a
higher quality of randomness to long, non-seeded runs which depend
on a very large number of random samples.

[^1]: https://gcc.gnu.org/onlinedocs/gcc-14.3.0/libstdc++/manual/manual/status.html#iso.2011.specific
[^2]: https://en.wikipedia.org/wiki/Linear_congruential_generator#Advantages_and_disadvantages
[^3]: https://peteroupc.github.io/hqprng.html#not-high-quality-prngs