Add regression testing of test outputs
This patch provides a bash script that enables the discovery of
unintentinoal changes to the output of the program due to changes
in the codebase when refactoring.

This patch moves the test.genotype.txt and test.location.txt files
to the tests directory for path consolidation. The documentation
referring to these files has subsequently been updated.