convert global config to struct
The primary purpose of this patch is to improve code readability and
understandability. This patch also opens the possibility of a config
file or higher-level language (Python/R) to be used to configure the
program without requiring a recompile.

For many, many reasons [^1][^2][^3], globals are considered a bad habit
especially in C++ code. When globals are used in a codebase, clarity and
intention are lost in the process. Clarity of knowing who is responsible
for setting and using the variable and intention of what the global is
supposed to signify and be a proxy for. Finally, the definition of
mutability is lost (const vs non-const) leaving the compiler to make
assumptions about whether a variable can change within a function.

In the case of SCAT3, a majority of the globals are configuration
variables set only during the menu processing and initialization phase
of the program. These globals are never changed afterwards, making them
an excellent candidate for passing to functions as a constant reference.
Now, the compiler can warn when a function accidentally tries to change
a configuration variable mid-run.

This patch pulls these globals in a new `struct Config` and adopts a
standardized documentation of menu syntax so that IDEs can provide
hover hints for variables on which flags will affect this variable
(and in some cases which indexed argument).

Additionally, this patch begins cleanup of not using namespace std
(see [^4] and [^5]) and conforming to a more consistent coding style.
Eventually, a future patch will provide a `clang-format` [^6] config to
provide automatic conformance for other code contributions through the
CLI and IDEs.

[^1]: https://www.codewithjason.com/global-variables/
[^2]: https://www.forrestthewoods.com/blog/global-variables-are-evil-and-unsafe/
[^3]: https://stackoverflow.com/q/484635
[^4]: https://stackoverflow.com/q/1452721
[^5]: https://builtin.com/articles/using-namespace-std
[^6]: https://clang.llvm.org/docs/ClangFormat.html